core/requirements.txt
Django==3.0
djangorestframework==3.13.1
asgiref==3.3.1
pytz==2021.1
sqlparse==0.4.1
psycopg2-binary==2.9.3
redis>=3.5
celery>=5.0
drf-yasg==1.20.0
